Adds GitHub authentication both as sign-in and for third-part authentication.
- Adds GitHub as an authentication provider for third-party authentication (ie: so we can kick of CI or something) https://backstage.io/docs/auth/github/provider
- Makes GitHub our authentication provider for sign-in https://backstage.io/docs/auth/#sign-in-configuration
- Adds GitHub as an entity provider for Backstage Users and groups https://backstage.io/docs/integrations/github/org/

GitHubOrgEntityProvider addition to the catalog is required
along with GITHUB_TOKEN that can read an orgs users so
the software catalog can populate with user and groups data.

Backstage's authentication resolvers need two pieces of information
https://backstage.io/docs/auth/identity-resolver#backstage-user-identity
One piece is a user entity reference which if you want to leverage built-in
auth resolvers such as `resolver: providers.github.resolvers.usernameMatchingUserEntityName()`
then you need to populate the software catalog with user/group data.

So this commit adds an entity provider for GitHubOrg easy-bull so that we can
leverage the build-in auth resolver and actually log into backstage via GitHub